Clustering funktioniert zumindest bei iris.csv nach meiner Ansicht vollstaendig:)

+++ distance.h
@@ -2,5 +2,5 @@
 #define _DISTANCE_H
 typedef enum distance {MANHATTAN, EUCLIDEAN} distance;
 
-double compute_distance(int d, const double x1[d], const double x2[d], distance dist);
+long double compute_distance(int d, const double x1[d], const double x2[d], distance dist);
 #endif
